PROM-3 Ajustadas criações de tabelas e funcionando 100% o database_reset

+++ promopredictor/src/data/data_cleaner.py
@@ -40,14 +40,19 @@ def clean_null_values(table_name, columns):
         logger.info(f"Limpeza na tabela '{table_name}': {affected_rows} linhas atualizadas onde {column} era NULL.")
 
 def remove_duplicates(table_name):
-    """Remove registros duplicados de uma tabela.
+    """
+    Remove registros duplicados de uma tabela. Utiliza chaves primárias para identificar duplicatas, caso disponíveis.
+    Caso contrário, usa todas as colunas da tabela.
 
     Args:
         table_name (str): Nome da tabela da qual os registros duplicados serão removidos.
     """
-    primary_key_columns = db_manager.get_primary_key_columns(table_name)
+    primary_key_columns = get_primary_key_columns(table_name)
     if primary_key_columns:
         primary_key_clause = ', '.join(primary_key_columns)
+        if not primary_key_clause:  # Se a cláusula de chaves primárias estiver vazia, evita executar uma query malformada
+            logger.error(f"Erro ao formar cláusula de chave primária para a tabela {table_name}.")
+            return
         query = f"""
             DELETE FROM {table_name}
             WHERE ({primary_key_clause}) NOT IN (
@@ -60,24 +65,33 @@ def remove_duplicates(table_name):
             );
         """
     else:
-        all_columns = db_manager.get_all_columns(table_name)
+        logger.info(f"Nenhuma chave primária encontrada, usando todas as colunas para a tabela {table_name}.")
+        all_columns = get_all_columns(table_name)
+        if not all_columns:  # Se não houver colunas, evita executar uma query malformada
+            logger.error(f"Nenhuma coluna encontrada para a tabela {table_name}. Impossível remover duplicatas.")
+            return
+        columns_clause = ', '.join(all_columns)
         query = f"""
             DELETE FROM {table_name}
-            WHERE ({', '.join(all_columns)}) NOT IN (
-                SELECT {', '.join(all_columns)}
+            WHERE ({columns_clause}) NOT IN (
+                SELECT {columns_clause}
                 FROM (
-                    SELECT {', '.join(all_columns)}, ROW_NUMBER() OVER (PARTITION BY {', '.join(all_columns)} ORDER BY {', '.join(all_columns)}) AS row_num
+                    SELECT {columns_clause}, ROW_NUMBER() OVER (PARTITION BY {columns_clause} ORDER BY {columns_clause}) AS row_num
                     FROM {table_name}
                 ) AS deduped
                 WHERE row_num = 1
             );
         """
 
-    success = db_manager.execute_query(query)
-    if success:
-        logger.info(f"Registros duplicados removidos da tabela '{table_name}'.")
-    else:
-        logger.error(f"Erro ao remover registros duplicados da tabela '{table_name}'.")
+    # Executando a query e logando o resultado
+    try:
+        success = db_manager.execute_query(query)
+        if success:
+            logger.info(f"Registros duplicados removidos da tabela '{table_name}'.")
+        else:
+            logger.error(f"Erro ao remover registros duplicados da tabela '{table_name}'. A query pode ter sido executada, mas sem sucesso na remoção.")
+    except Exception as e:
+        logger.error(f"Erro ao executar a remoção de duplicatas na tabela {table_name}: {e}")
 
 def remove_invalid_records(table_name, conditions):
     """
@@ -89,43 +103,47 @@ def remove_invalid_records(table_name, conditions):
     for condition in conditions:
         delete_data(table_name, condition)
 
-def get_primary_key_columns(self, table_name):
+def get_primary_key_columns(table_name):
     """
     Obtém os nomes das colunas que compõem a chave primária de uma tabela.
-
     Args:
         table_name (str): Nome da tabela.
-
     Returns:
         list: Lista com os nomes das colunas que compõem a chave primária.
     """
-    if self.use_sqlalchemy:
-        # Implementação para SQLAlchemy
-        pass
-    else:
-        query = f"SHOW KEYS FROM {table_name} WHERE Key_name = 'PRIMARY'"
-        result = self.execute_query(query)
-        primary_key_columns = [row[4] for row in result]
+    logger.info(f"Tentando obter chaves primárias da tabela {table_name}")
+    query = f"SHOW KEYS FROM {table_name} WHERE Key_name = 'PRIMARY'"
+    try:
+        result = db_manager.execute_query(query)
+        if result is None or len(result) == 0:
+            logger.error(f"Erro ao obter chaves primárias para a tabela {table_name} ou nenhuma chave primária encontrada.")
+            return []
+        primary_key_columns = [row[4] for row in result]  # Ajustar o índice conforme a estrutura real do resultado.
         return primary_key_columns
+    except Exception as e:
+        logger.error(f"Erro ao processar chaves primárias para a tabela {table_name}: {e}")
+        return []
 
-def get_all_columns(self, table_name):
+def get_all_columns(table_name):
     """
     Obtém os nomes de todas as colunas de uma tabela.
-
     Args:
         table_name (str): Nome da tabela.
-
     Returns:
         list: Lista com os nomes de todas as colunas da tabela.
     """
-    if self.use_sqlalchemy:
-        # Implementação para SQLAlchemy
-        pass
-    else:
-        query = f"DESCRIBE {table_name}"
-        result = self.execute_query(query)
+    logger.info(f"Tentando obter todas as colunas da tabela {table_name}")
+    query = f"DESCRIBE {table_name}"
+    try:
+        result = db_manager.execute_query(query)
+        if result is None or len(result) == 0:
+            logger.error(f"Erro ao descrever a tabela {table_name} ou nenhuma coluna encontrada.")
+            return []
         column_names = [row[0] for row in result]
         return column_names
+    except Exception as e:
+        logger.error(f"Erro ao tentar obter colunas da tabela {table_name}: {e}")
+        return []
 
 def standardize_formatting(table_name, formatting_rules):
     """
